Use glogging module for logs
